Fix SD restart error in exe file
-- This commit fixes SD restart error in exe file by creating
   variants.json in CWD instead of a relative path.
