+++ apps/stable_diffusion/src/utils/resources.py
@@ -35,28 +35,3 @@ base_models = get_json_file("resources/base_model.json")
 
 # Contains optimization flags for different models.
 opt_flags = get_json_file("resources/opt_flags.json")
-
-
-# `fetch_and_update_base_model_id` is a resource utility function which
-# helps maintaining mapping of the model to run with its base model.
-# If `base_model` is "", then this function tries to fetch the base model
-# info for the `model_to_run`.
-def fetch_and_update_base_model_id(model_to_run, base_model=""):
-    path = "resources/variants.json"
-    loc_json = resource_path(path)
-    data = {model_to_run: base_model}
-    json_data = {}
-    if os.path.exists(loc_json):
-        with open(loc_json, "r", encoding="utf-8") as jsonFile:
-            json_data = json.load(jsonFile)
-            # Return with base_model's info if base_model is "".
-            if base_model == "":
-                if model_to_run in json_data:
-                    base_model = json_data[model_to_run]
-                return base_model
-    elif base_model == "":
-        return base_model
-    # Update JSON data to contain an entry mapping model_to_run with base_model.
-    json_data.update(data)
-    with open(loc_json, "w", encoding="utf-8") as jsonFile:
-        json.dump(json_data, jsonFile)
